refactor: refactor sagas
