--- src/redux/sagas/error-tip.ts
@@ -1,22 +1,0 @@
-import { ForkEffect } from 'redux-saga/effects';
-// import { AnyAction } from 'redux';
-
-export { ForkEffect };
-
-// export default function* errorTipSaga() {
-//   // yield takeEvery('LOGIN_FAILED', function*(action: AnyAction) {
-//   //   yield put({ type: 'SET_LOGIN_ERROR', payload: `登录失败：${action.payload}` });
-//   // });
-//   // yield takeEvery('REGISTER_FAILED', function*(action: AnyAction) {
-//   //   yield put({ type: 'SET_REGISTER_ERROR', payload: `注册失败：${action.payload}` });
-//   // });
-//   // yield takeEvery('REGISTER_SMS_FAILED', function*(action: AnyAction) {
-//   //   yield put({ type: 'SET_REGISTER_ERROR', payload: `发送短信失败：${action.payload}` });
-//   // });
-//   // yield takeEvery('RESET_PWD_SMS_FAILED', function*(action: AnyAction) {
-//   //   yield put({ type: 'SET_RESET_PWD_ERROR', payload: `发送短信失败：${action.payload}` });
-//   // });
-//   // yield takeEvery('NEW_TEAM_SUBMIT_FAILED', function*(action: AnyAction) {
-//   //   yield put({ type: 'SET_NEW_TEAM_ERROR', payload: `创建队伍失败：${action.payload}` });
-//   // });
-// }
